правки после изменения БД

+++ Quiz/Repositories/Implementations/QuestionRepository.cs
@@ -15,15 +15,30 @@ public class QuestionRepository : IQuestionRepository
 
     public async Task<Question?> GetByIdAsync(int id)
     {
+        return await _context.Questions.FindAsync(id);
+    }
+
+    public async Task<Question?> GetByIdWithOptionsAsync(int id)
+    {
         return await _context.Questions
+            .Include(q => q.Options)
             .FirstOrDefaultAsync(q => q.Id == id);
     }
 
     public async Task<IEnumerable<Question>> GetQuestionsByQuizAsync(int quizId)
     {
         return await _context.Questions
-           .Where(q => q.QuizId == quizId)
-           .ToListAsync();
+            .Where(q => q.QuizId == quizId)
+            .ToListAsync();
+    }
+
+    public async Task<IEnumerable<Question>> GetQuestionsWithOptionsByQuizAsync(int quizId)
+    {
+        return await _context.Questions
+            .Where(q => q.QuizId == quizId)
+            .Include(q => q.Options)
+            .OrderBy(q => q.Id)
+            .ToListAsync();
     }
 
     public async Task AddAsync(Question question)
@@ -38,9 +53,14 @@ public class QuestionRepository : IQuestionRepository
         await _context.SaveChangesAsync();
     }
 
-    public async Task DeleteAsync(Question question)
+    public async Task DeleteAsync(int id)
     {
-        _context.Questions.Remove(question);
-        await _context.SaveChangesAsync();
+        var question = await _context.Questions.FindAsync(id);
+
+        if (question != null)
+        {
+            _context.Questions.Remove(question);
+            await _context.SaveChangesAsync();
+        }
     }
 }
